# api-docs/API.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'Online-LMS APIs'
  description: 'List of REST APIs Documentation for Online Learning Management System'
  contact:
    name: Palden Sherpa
    email: sherpalden369@gmail.com
servers: 
  - url: http://localhost:5000
    description: Local Server
    
  - url: https://aiex-olms.herokuapp.com
    description: Heroku Server
paths:
...
  /api/admin/news/add-main-category:
    post:
      tags:
        - News Category APIs - Admin
      description: Add main news category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              type: object
              properties:
                categoryName:
                  type: string
                  example: Science and myths
                  
      responses:
        201:
          description: Response on successful addition of main news category
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "News main category addition successful"
  /api/admin/news/add-sub-category:
    post:
      tags:
        - News Category APIs - Admin
      description: Add sub category for news
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              type: object
              properties:
                categoryName:
                  type: string
                  example: Animal
                parentID:
                  type: string
                  example: adfadf7987sdfs98798sdf
                  
      responses:
        201:
          description: Response on successful addition of sub category
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "News sub category addition successful"
  /api/admin/news/rename-category:
    put:
      tags:
        - News Category APIs - Admin
      description: Rename a category for news
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              type: object
              properties:
                newName:
                  type: string
                  example: Animal
                categoryID:
                  type: string
                  example: adfadf7987sdfs98798sdf
                  
      responses:
        201:
          description: Response on successful rename of a category
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Category rename successful"
  /api/admin/news/delete-category/{categoryID}:
    delete:
      tags:
        - News Category APIs - Admin
      description: Deletes all the sub categories and news associated with them including the category itself.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: categoryID
      responses:
        201:
          description: Response on successful deletion of a category
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Category delete successful"
  /api/admin/news/all-categories:
    get:
      tags:
        - News Category APIs - Admin
      description: Get all category list
      security:
        - bearerAuth: []
      responses:
        201:
          description: Response on successful retrieval of all category list
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "All category list retrieval successful"
  /api/admin/news/all-category-tree:
    get:
      tags:
        - News Category APIs - Admin
      description: Get all category list in tree view
      security:
        - bearerAuth: []
      responses:
        201:
          description: Response on successful retrieval of all category list in tree view
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "All category list retrieval in tree view successful"


#******
#******
#******
#******
#******
#******
  /api/admin/news/upload-images:
    post:
      tags:
        - News APIs - Admin
      description: Upload news images
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content: 
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
              
      responses:
        200:
          description: Response on successful upload of news images
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: News Images Upload Successful
  /api/admin/news:
    post:
      tags:
        - News APIs - Admin
      description: Create a news post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content: 
          multipart/form-data:
            schema:
              type: object
              properties:
                # files:
                #   type: array
                #   items:
                #     type: string
                #     format: binary
                images:
                  type: array
                  example: ["img1.jpg", "img2.jpg"]
                title:
                  type: string
                  example: Nepal is a agricultural country but we import agro products from India.
                categoryID:
                  type: string
                  example: 345dght56757ghjg678
                description:
                  type: string
                  example: News description here
                youtubeLink:
                  type: string
                  example: https://youtu.be/QNInfBuBcOA
                publishingAt:
                  type: string
                  example: 2020-20-1
      responses:
        200:
          description: Response on successful creation of news post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: News Post Creation Successful
  /api/admin/news/{newsID}:
    get:
      tags:
        - News APIs - Admin
      description: Get the details data of the particular news post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: newsID
      responses:
        200:
          description: Response on successful retrieval of detail data of the particular news post
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: News post details retrieval successful

    put:
      tags:
        - News APIs - Admin
      description: Update a news post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: newsID
      requestBody:
        required: true
        content: 
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                title:
                  type: string
                  example: Nepal is a agricultural country but we import agro products from India.
                categoryID:
                  type: string
                  example: 345dght56757ghjg678
                description:
                  type: string
                  example: News description here
                metaTags:
                  type: array
                  example: ["tag-1", "tag-2", "tag-3"]
                youtubeLink:
                  type: string
                  example: https://youtu.be/QNInfBuBcOA
                imagesToBeDeleted:
                  type: array
                  example: ["img-1.jpg", "img-2.jpg"]
                publishingAt:
                  type: string
                  example: 2020-20-1
      responses:
        200:
          description: Response on successful udpate of news post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: News Update Successful

    delete:
      tags:
        - News APIs - Admin
      description: Delete a news post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: newsID
      responses:
        200:
          description: Response on successful deletion of news post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: News delete Successful
  /api/admin/news/all-news/{categoryID}:
    get:
      tags:
        - News APIs - Admin
      description: Get all the news of the particular categoryID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: categoryID
      responses:
        200:
          description: Response on successful retrieval of news of particular categoryID
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: News posts of particular categoryID retrieval successful

    delete:
      tags:
        - News APIs - Admin
      description: Delete all news of particular categroyID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: categoryID
      responses:
        200:
          description: Response on successful deletion of news of particular categoryID
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: News of particular category delete Successful
  /api/admin/news/set-thumbnail/{newsID}:
    put:
      tags:
        - News APIs - Admin
      description: Set a thumbnail image for a news post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: newsID
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              type: object
              properties:
                thumbnailImage:
                  type: string
                  example: img-1.jpg
      responses:
        200:
          description: Response on successful set of thumbnail image for news post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: News thumbnail image set Successful
  /api/admin/news/search:
    get:
      tags:
        - News APIs - Admin
      description: Search a news post
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: searchKey
        - in: query
          name: skips
        - in: query
          name: limit
     
      responses:
        200:
          description: Response on successful Search of news post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: news Post Search Successful

#******
#******
#******
#******
#******
#******
...
  /api/user/news/{newsID}:
    get:
      tags:
        - News APIs - User
      description: Get the details data of the particular news post
      parameters:
        - in: path
          name: newsID
      responses:
        200:
          description: Response on successful retrieval of detail data of the particular news post
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: News post details retrieval successful
  /api/user/news/all-news/{categoryID}:
    get:
      tags:
        - News APIs - User
      description: Get all the news of the particular categoryID
      parameters:
        - in: path
          name: categoryID
      responses:
        200:
          description: Response on successful retrieval of news of particular categoryID
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: News posts of particular categoryID retrieval successful

#******
#******
#******
#******
#******
#******
...
components:
  securitySchemes:
    bearerAuth:           
      type: http
      scheme: bearer
      bearerFormat: JWT  